requirements.txt
google-api-python-client==2.116.0
google-auth-oauthlib==1.2.0
beautifulsoup4==4.12.3
pyyaml==6.0.1
